[Update] - Split code into components + API for Formula page
