+++ src/pages/formula/components/index.ts
@@ -1,0 +1,6 @@
+export * from "@/pages/formula/components/FormulaDetailTab";
+export * from "@/pages/formula/components/ApprovedHistoryTab";
+export * from "@/pages/formula/components/ModalConfirm";
+export * from "@/pages/formula/components/ModalFormula";
+export * from "@/pages/formula/components/ModalMaterial";
+export * from "@/pages/formula/components/ModalNote";
